+++ src/mailman/pipelines/builtin.py
@@ -61,6 +61,8 @@ class PostingPipeline(BasePipeline):
         # All decoration is now done in delivery.
         # 'decorate',
         'dmarc',
-        'arc-sign',
+        # Message decoration in delivery can break an arc signature, so sign
+        # in delivery after decorating.
+        # 'arc-sign',
         'to-outgoing',
         )